"simplifying" docker-compose
Not being able to have signed TLS certs makes things more complicated to
test, because one need to validate on every endpoints the certificate
before being able to launch the app.

We are going to run the app from a closed env (no public IP, so no LE
cert can be generated).